Scanner prototype 1
The scanner function is now working. When parse is called and a file name is passed in, a list is created from the file. Then the scanner is called with that list being passed to it. The scanner first changes the datatype of every datum in the list to be of data type string. Then the scanner goes through each item in the list and reassigns it as a token for the parser.